agent1/requirements.txt
# Multi-Agent Commercial District Analysis System
# Core Dependencies

# LLM & Agent Framework
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.62
langgraph-checkpoint==0.0.4
openai==1.10.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2

# Web Scraping
selenium==4.16.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# Computer Vision
opencv-python==4.9.0.80
pillow==10.2.0

# Dashboard & Visualization
streamlit==1.30.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# MCP (Model Context Protocol)
mcp[cli]==0.4.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3

# Development
pytest==7.4.4
black==23.12.1
flake8==7.0.0

